+++ spec/support/suspenders.rb
@@ -11,50 +11,90 @@ module SuspendersTestHelpers
 
   def run_suspenders(arguments = nil)
     arguments = "--path=#{root_path} #{arguments}"
-    Dir.chdir(tmp_path) do
-      Bundler.with_clean_env do
-        add_fakes_to_path
-        `
-          #{suspenders_bin} #{APP_NAME} #{arguments}
-        `
-        Dir.chdir(APP_NAME) do
-          with_env("HOME", tmp_path) do
-            `git add .`
-            `git commit -m 'Initial commit'`
-          end
+    run_in_tmp do
+      add_fakes_to_path
+
+      with_revision_for_honeybadger do
+        debug `#{suspenders_bin} #{APP_NAME} #{arguments}`
+      end
+
+      Dir.chdir(APP_NAME) do
+        with_env("HOME", tmp_path) do
+          debug `git add .`
+          debug `git commit -m 'Initial commit'`
         end
+      end
+    end
+  end
+
+  def with_app
+    drop_dummy_database
+    remove_project_directory
+    rails_new
+    setup_app_dependencies
+
+    yield
+  end
+
+  def rails_new
+    run_in_tmp do
+      add_fakes_to_path
+
+      with_revision_for_honeybadger do
+        debug `#{system_rails_bin} new #{APP_NAME}`
+      end
+
+      Dir.chdir(APP_NAME) do
+        File.open("Gemfile", "a") do |file|
+          file.puts %{gem "suspenders", path: #{root_path.inspect}}
+        end
+
+        with_env("HOME", tmp_path) do
+          debug `git add .`
+          debug `git commit -m 'Initial commit'`
+        end
+      end
+    end
+  end
+
+  def generate(generator)
+    run_in_project do
+      with_revision_for_honeybadger do
+        debug `bin/spring stop`
+        debug `#{project_rails_bin} generate #{generator}`
+      end
+    end
+  end
+
+  def destroy(generator)
+    run_in_project do
+      with_revision_for_honeybadger do
+        `bin/spring stop`
+        `#{project_rails_bin} destroy #{generator}`
       end
     end
   end
 
   def suspenders_help_command
-    Dir.chdir(tmp_path) do
-      Bundler.with_clean_env do
-        `
-          #{suspenders_bin} -h
-        `
-      end
+    run_in_tmp do
+      debug `#{suspenders_bin} -h`
     end
   end
 
   def setup_app_dependencies
-    if File.exist?(project_path)
-      Dir.chdir(project_path) do
-        Bundler.with_clean_env do
-          `bundle check || bundle install`
-        end
-      end
+    run_in_project do
+      debug `bundle check || bundle install`
     end
+  rescue Errno::ENOENT
+    # The project_path might not exist, in which case we can skip this.
   end
 
   def drop_dummy_database
-    if File.exist?(project_path)
-      Dir.chdir(project_path) do
-        Bundler.with_clean_env do
-          `rails db:drop`
-        end
-      end
+    run_in_project do
+      debug `#{project_rails_bin} db:drop 2>&1`
     end
+  rescue Errno::ENOENT
+    # The project_path might not exist, in which case we can skip this.
   end
 
   def add_fakes_to_path
@@ -79,6 +119,14 @@ module SuspendersTestHelpers
     File.join(root_path, 'bin', 'welaika-suspenders')
   end
 
+  def system_rails_bin
+    "rails"
+  end
+
+  def project_rails_bin
+    "bin/rails"
+  end
+
   def support_bin
     File.join(root_path, "spec", "fakes", "bin")
   end
@@ -88,12 +136,47 @@ module SuspendersTestHelpers
   end
 
   def with_env(name, new_value)
+    had_key = ENV.has_key?(name)
     prior = ENV[name]
     ENV[name] = new_value.to_s
 
     yield
 
   ensure
-    ENV[name] = prior
+    ENV.delete(name)
+
+    if had_key
+      ENV[name] = prior
+    end
+  end
+
+  def with_revision_for_honeybadger
+    with_env("HEROKU_SLUG_COMMIT", 1) do
+      yield
+    end
+  end
+
+  def run_in_tmp
+    Dir.chdir(tmp_path) do
+      Bundler.with_clean_env do
+        yield
+      end
+    end
+  end
+
+  def run_in_project
+    Dir.chdir(project_path) do
+      Bundler.with_clean_env do
+        yield
+      end
+    end
+  end
+
+  def debug(output)
+    if ENV["DEBUG"]
+      warn output
+    end
+
+    output
   end
 end